+++ govtool/backend/src/VVA/Config.hs
@@ -26,6 +26,7 @@ module VVA.Config
     , vvaConfigToText
     , getMetadataValidationHost
     , getMetadataValidationPort
+    , getWebsocketLifetimeSeconds
     ) where
 
 import           Conferer
@@ -88,6 +89,10 @@ data VVAConfigInternal
       , vVAConfigInternalMetadataValidationPort :: Int
         -- | Maximum number of concurrent metadata requests
       , vVAConfigInternalMetadataValidationMaxConcurrentRequests :: Int
+        -- | Redis config
+      , vVAConfigInternalRedisConfig :: RedisInternalConfig
+        -- | WebSocket lifetime in seconds
+      , vVAConfigInternalWebsocketLifetimeSeconds :: Int
       }
   deriving (FromConfig, Generic, Show)
 
@@ -102,7 +107,9 @@ instance DefaultConfig VVAConfigInternal where
         vVAConfigInternalSentryEnv = "development",
         vVAConfigInternalMetadataValidationHost = "localhost",
         vVAConfigInternalMetadataValidationPort = 3001,
-        vVAConfigInternalMetadataValidationMaxConcurrentRequests = 10
+        vVAConfigInternalMetadataValidationMaxConcurrentRequests = 10,
+        vVAConfigInternalRedisConfig = RedisInternalConfig "localhost" 6379 Nothing,
+        vVAConfigInternalWebsocketLifetimeSeconds = 60 * 1
       }
 
 -- | DEX configuration.
@@ -126,6 +133,10 @@ data VVAConfig
       , metadataValidationPort :: Int
         -- | Maximum number of concurrent metadata requests
       , metadataValidationMaxConcurrentRequests :: Int
+        -- | Redis config
+      , redisConfig :: RedisConfig
+        -- | WebSocket lifetime in seconds
+      , websocketLifetimeSeconds :: Int
       }
   deriving (Generic, Show, ToJSON)
 
@@ -169,7 +180,13 @@ convertConfig VVAConfigInternal {..} =
       sentryEnv = vVAConfigInternalSentryEnv,
       metadataValidationHost = vVAConfigInternalMetadataValidationHost,
       metadataValidationPort = vVAConfigInternalMetadataValidationPort,
-      metadataValidationMaxConcurrentRequests = vVAConfigInternalMetadataValidationMaxConcurrentRequests
+      metadataValidationMaxConcurrentRequests = vVAConfigInternalMetadataValidationMaxConcurrentRequests,
+      redisConfig = RedisConfig
+        { redisHost = redisInternalConfigHost $ vVAConfigInternalRedisConfig,
+          redisPort = redisInternalConfigPort $ vVAConfigInternalRedisConfig,
+          redisPassword = redisInternalConfigPassword $ vVAConfigInternalRedisConfig
+        },
+      websocketLifetimeSeconds = vVAConfigInternalWebsocketLifetimeSeconds
     }
 
 -- | Load configuration from a file specified on the command line.  Load from
@@ -219,3 +236,9 @@ getMetadataValidationPort ::
   (Has VVAConfig r, MonadReader r m) =>
   m Int
 getMetadataValidationPort = asks (metadataValidationPort . getter)
+
+-- | Access websocket lifetime in seconds
+getWebsocketLifetimeSeconds ::
+  (Has VVAConfig r, MonadReader r m) =>
+  m Int
+getWebsocketLifetimeSeconds = asks (websocketLifetimeSeconds . getter)